Add files via upload
Use the Titanic dataset with help of R Studio to build a model that predicts whether a passenger on the Titanic survived or
not. This is a classic beginner project with readily available data.
The dataset typically used for this project contains information about individual passengers, such as
their age, gender, ticket class, fare, cabin, and whether or not they survived.